app/requirements.txt
# Framework
starlette == 0.13.4
uvicorn == 0.11.5

# Security
passlib==1.7.2
PyJWT==1.7.1

# Database
tortoise-orm == 0.16.13
asyncpg == 0.20.1

# GraphQL
ariadne ==0.11.0